[TACHYON-1232] Move tachyon.LocalFilesystemCluster to tachyon.underfs.LocalFilesystemCluster
