+++ minicluster/src/main/java/tachyon/underfs/UnderFileSystemCluster.java
@@ -22,7 +22,7 @@ import org.apache.commons.lang3.StringUtils;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Throwables;
 
-import tachyon.LocalFilesystemCluster;
+import tachyon.underfs.LocalFilesystemCluster;
 import tachyon.TachyonURI;
 import tachyon.conf.TachyonConf;
 import tachyon.util.io.PathUtils;
